chore: move classes into new appmap package
